add dom updates file and transfered display cards to new file and made clear card grid a helper function
